Switch to Utilities-OG for messages

+++ src/main/java/com/tcoded/playerbountiesplus/event/BountySetEvent.java
@@ -41,7 +41,7 @@ public class BountySetEvent extends Event implements Cancellable {
 
     }
 
-    public float getAmount() {
+    public double getAmount() {
 
         return amount;
